Move switch account functionality to accept invitation card
